improvement: add a new service function for reading the file
