+++ bigbluebutton-html5/imports/ui/components/video-preview/virtual-background/service.js
@@ -1,5 +1,9 @@
+import _ from 'lodash';
 import logger from '/imports/startup/client/logger';
 import { notify } from '/imports/ui/services/notification';
+
+const MIME_TYPES_ALLOWED = ['image/png', 'image/jpeg'];
+const MAX_FILE_SIZE = 5000; // KBytes
 
 const withObjectStore = ({
   onError,
@@ -92,8 +96,44 @@ const del = (key) => {
   });
 };
 
+const parseFilename = (filename = '') => {
+  const substrings = filename.split('.');
+  substrings.pop();
+  const filenameWithoutExtension = substrings.join('');
+  return filenameWithoutExtension;
+};
+
+const readFile = (file, onSuccess, onError) => {
+  const { name, size, type } = file;
+  const sizeInKB = size / 1024;
+
+  if (sizeInKB > MAX_FILE_SIZE) {
+    return onError(new Error('Maximum file size exceeded.'));
+  }
+
+  if (!MIME_TYPES_ALLOWED.includes(type)) {
+    return onError(new Error('File type not allowed.'));
+  }
+
+  const filenameWithoutExtension = parseFilename(name);
+  const reader = new FileReader();
+
+  reader.onload = function (e) {
+    const background = {
+      filename: filenameWithoutExtension,
+      data: e.target.result,
+      uniqueId: _.uniqueId(),
+    };
+    onSuccess(background);
+  }
+  reader.readAsDataURL(file);
+};
+
 export default {
   load,
   save,
   del,
+  readFile,
+  MIME_TYPES_ALLOWED,
+  MAX_FILE_SIZE,
 };
